videochat/src/py/requirements.txt
WebOb
Paste
webapp2==3.0.0b1
jinja2
google-api-python-client
appengine-python-standard>=0.1.1
redis
oauth2client
Django==3.2.5
djangorestframework